+++ src/aioverse/errors/__init__.py
@@ -1,5 +1,7 @@
 from .ResponseCodeError import ResponseCodeError
+from .sse_parse_error import SSEParseError
 
 __all__ = [
-	"ResponseCodeError"
+	"ResponseCodeError",
+	"SSEParseError",
 ]